create new component ModalList to configure the columns of the tables, implement component in PotentialDuplicatesList
